+++ app/context/provier.js
@@ -3,5 +3,5 @@
 import { SessionProvider } from "next-auth/react";
 
 export default function Provider({ children, session }) {
-  return <SessionProvider session={session}>{children}</SessionProvider>;
+  return <SessionProvider session={session}>{children}  </SessionProvider>;
 }
