+++ app/utils/handleButton.js
@@ -1,7 +1,7 @@
  const handleButton = (data, setData, field) => {
     return (
       <>
-        <div className="flex items-center justify-between mt-5">
+        <div className="flex items-center justify-between mt-2">
           <button
             type="button"
             onClick={() => setData(!data)}
